backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.4
pydantic-settings==2.3.4
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
pytest==8.2.2
httpx==0.27.0
python-dotenv==1.0.1
alembic==1.13.1
